include/types.h: move client_t to separate header
