+++ src/crimson/CMakeLists.txt
@@ -111,6 +111,7 @@ add_library(crimson-common STATIC
   ${PROJECT_SOURCE_DIR}/src/global/pidfile.cc
   ${PROJECT_SOURCE_DIR}/src/include/filepath.cc
   ${PROJECT_SOURCE_DIR}/src/include/object.cc
+  ${PROJECT_SOURCE_DIR}/src/include/client_t.cc
   ${PROJECT_SOURCE_DIR}/src/include/types.cc
   ${PROJECT_SOURCE_DIR}/src/include/errorcode32.cc
   ${PROJECT_SOURCE_DIR}/src/include/shard_id.cc
