+++ src/include/cephfs/types.h
@@ -24,6 +24,7 @@
 #include <variant>
 #include <vector>
 
+#include "include/client_t.h"
 #include "include/compact_set.h"
 #include "include/encoding.h"
 #include "include/fs_types.h"
