+++ src/mds/Capability.h
@@ -19,6 +19,7 @@
 
 #include "include/buffer_fwd.h"
 #include "include/ceph_fs.h" // for CEPH_CAP_*
+#include "include/client_t.h"
 #include "include/counter.h"
 #include "include/mempool.h"
 #include "include/object.h" // for snapid_t
