+++ src/osd/PrimaryLogPG.h
@@ -19,7 +19,7 @@
 
 #include <boost/tuple/tuple.hpp>
 #include "include/ceph_assert.h"
-#include "include/types.h" // for client_t
+#include "include/client_t.h"
 #include "DynamicPerfStats.h"
 #include "OSD.h"
 #include "PG.h"
